Refactor import order in main.ts and connector.ts
