+++ package.json
@@ -9,14 +9,18 @@
   "dependencies": {
     "@colors/colors": "^1.6.0",
     "@types/needle": "^2.5.3",
-    "@types/node": "^18.11.9",
-    "@types/yargs": "^17.0.32",
     "dotenv": "^16.0.3",
     "env-var": "^7.4.1",
+    "fs-extra": "^11.2.0",
     "needle": "^3.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.4.5",
     "winston": "^3.13.0",
     "yargs": "^17.7.2"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^11.0.4",
+    "@types/node": "^20.12.11",
+    "@types/yargs": "^17.0.32"
   }
 }
